Created an Animation Controller for the character
Created a new animation controller for the character in order to animate the player. Animated the character's running and falling animationThe animations used come from the Supercyan Character Pack Free Sample Asset.

Created a new trigger parameter, is falling. Created a condition in the transition to activate the falling animation when the character isFalling parameter is true.