Modified the CharacterMovement script
Added 2 new properties.
rayStart, a Transform object used to set a position for a raycast to start from the character's belly. This will be used to trigger Raycast under the character and check if there's no physical object for the character to run on.

anim, an animation object used to Set the Trigger parameter whenever the condition is met (Meaning the character is falling).

Added code to the Update method to check if the raycast hits infinity downward, meaning no physical floor, meaning character is falling. Skadoosh

If the character is falling, we set the Trigger variable in the animation controller