Modified the CharacterMovement Script
Added and initialized a GameManager object in order to access the gameStarted boolean and method "StartGame" that sets this variable to true.

Added logic to check if game has started or not.

Added new animation "Idle" to the Animation controller for when the game has not yet been started.

Set Idle animation as default animation.

Added new trigger parameter gameStarted to the animation controller. This trigger parameter is set in the CharacterMovement script when the game has been started.

Added new condition to the transition from Idle to Run. Condition parameter is the "gameStarted" trigger parameter.