Multiple changes in Main Scene
Created new prefab using the ring from the Imported SimpleCollectiblesPack asset.

Added rings to the stage

Created Empty Object Rings to store all the collectible rings

Created UI text element to display the current score of the game. The score increments when rings are collected.

Added scoreText object to the GameManager script in order for the scoreText to update when a ring is collected.